+++ hpOneView/resources/networking/logical_interconnect_groups.py
@@ -44,14 +44,15 @@ class LogicalInterconnectGroups(object):
     DEFAULT_VALUES = {
         '200': {"type": "logical-interconnect-groupV3"},
         '300': {"type": "logical-interconnect-groupV300"},
-        '500': {"type": "logical-interconnect-groupV300"}
+        '500': {"type": "logical-interconnect-groupV300"},
+        '600': {"type": "logical-interconnect-groupV4"}
     }
 
     def __init__(self, con):
         self._connection = con
         self._client = ResourceClient(con, self.URI)
 
-    def get_all(self, start=0, count=-1, filter='', sort=''):
+    def get_all(self, start=0, count=-1, filter='', sort='', scope_uris=''):
         """
         Gets a list of logical interconnect groups based on optional sorting and filtering and is constrained by start
         and count parameters.
@@ -70,11 +71,14 @@ class LogicalInterconnectGroups(object):
             sort:
                 The sort order of the returned data set. By default, the sort order is based
                 on create time with the oldest entry first.
+            scope_uris:
+                An expression to restrict the resources returned according to the scopes to
+                which they are assigned.
 
         Returns:
             list: A list of logical interconnect groups.
         """
-        return self._client.get_all(start, count, filter=filter, sort=sort)
+        return self._client.get_all(start, count, filter=filter, sort=sort, scope_uris=scope_uris)
 
     def get(self, id_or_uri):
         """