+++ hpOneView/resources/resource.py
@@ -96,7 +96,7 @@ class ResourceClient(object):
         self._uri = uri
         self._task_monitor = TaskMonitor(con)
 
-    def build_query_uri(self, start=0, count=-1, filter='', query='', sort='', view='', fields='', uri=None):
+    def build_query_uri(self, start=0, count=-1, filter='', query='', sort='', view='', fields='', uri=None, scope_uris=''):
         """
         Builds the URI given the parameters.
 
@@ -135,6 +135,9 @@ class ResourceClient(object):
                 Name of the fields.
             uri:
                 A specific URI (optional)
+            scope_uris:
+                An expression to restrict the resources returned according to the scopes to
+                which they are assigned.
 
         Returns:
             uri: The complete uri
@@ -155,17 +158,20 @@ class ResourceClient(object):
         if fields:
             fields = "&fields=" + quote(fields)
 
+        if scope_uris:
+            scope_uris = "&scope_uris=" + quote(scope_uris)
+
         path = uri if uri else self._uri
         self.__validate_resource_uri(path)
 
         symbol = '?' if '?' not in path else '&'
 
-        uri = "{0}{1}start={2}&count={3}{4}{5}{6}{7}{8}".format(path, symbol, start, count, filter, query, sort,
-                                                                view, fields)
+        uri = "{0}{1}start={2}&count={3}{4}{5}{6}{7}{8}{9}".format(path, symbol, start, count, filter, query, sort,
+                                                                   view, fields, scope_uris)
 
         return uri
 
-    def get_all(self, start=0, count=-1, filter='', query='', sort='', view='', fields='', uri=None):
+    def get_all(self, start=0, count=-1, filter='', query='', sort='', view='', fields='', uri=None, scope_uris=''):
         """
         Gets all items according with the given arguments.
 
@@ -192,13 +198,16 @@ class ResourceClient(object):
                 Name of the fields.
             uri:
                 A specific URI (optional)
+            scope_uris:
+                An expression to restrict the resources returned according to the scopes to
+                which they are assigned.
 
         Returns:
             list: A list of items matching the specified filter.
         """
 
         uri = self.build_query_uri(start=start, count=count, filter=filter,
-                                   query=query, sort=sort, view=view, fields=fields, uri=uri)
+                                   query=query, sort=sort, view=view, fields=fields, uri=uri, scope_uris=scope_uris)
 
         logger.debug('Getting all resources with uri: {0}'.format(uri))