+++ CHANGELOG.md
@@ -1,7 +1,14 @@
-# 4.4.1 (Unreleased)
+# 4.5.0 (Unreleased)
 #### Notes
 Added the capability to set a connection timeout when connecting to the HPE OneView Appliance
+Extends support of the SDK to OneView Rest API version 600 (OneView v4.0).
 
+#### Features supported with current release:
+- Connection template
+- FC network
+- Interconnect type
+- Logical interconnect group
+- SAS logical interconnect group
 
 # 4.4.0
 #### Notes